requirements.txt
Django==4.2.7
qrcode[pil]==7.4.2
Pillow==10.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
whitenoise==6.5.0
psycopg2-binary==2.9.9
dj-database-url==2.0.0
